refactor(myinfo): split JWKS and DPoP handling out of the FAPI client
myinfo.fapi.client.ts had grown three unrelated jobs: OIDC discovery and
client configuration, RP keyset loading, and the DPoP proof helpers. Only
the first is the client's own concern, and the other two are what the
routes and the service actually reach for, so importing the client to get
at a key import made the module graph read backwards.

Splits the keyset code into myinfo.fapi.jwks.ts (its natural home: the
JWKS route already only wants getPublicJwks) and the DPoP helpers into
myinfo.fapi.dpop.ts, leaving the client with discovery alone. Key
generation now goes through generateDpopKey rather than the service
hand-rolling randomDPoPKeyPair plus exportPrivateJwk.

Pure move, no behaviour change. myinfo.fapi.client.spec.ts is renamed to
myinfo.fapi.jwks.spec.ts to follow the code it covers.
